+++ CHANGELOG.md
@@ -11,11 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - support for external tools
 - dedicated log (per conversation) of LLM request/responses
+- added remove_files tool
 
 ### Changed
 
 - docs shuffle
-- updated deps
+- updated deps (including Anthropic SDK with GA prompt caching)
 
 ### Fixed